Cache circles in LineChartRenderer instead of recreating them (#1897)
IDataSet now has a UserData object which can store caches of data that may be expensive to calculate on-the-fly.  In this case, LineChartRenderer defines a UserData object to pre-draw and cache the circles that would be drawn inside of drawCircles().  Since canvas.drawBitmap() offers significant savings over the previous canvas.drawCircle()/canvas.drawPath(), we reduce the worst case rendering times by up to 40%.  This savings is noticeable in the RealtimeLineChartActivity.  Said activity is updated to contain many different colors, to ensure that the bitmap caching properly and correctly paints the circles as expected.