LineChartRenderer takes responsibility for caching circle bitmaps (#1897)
Upon further consideration, IDataSet containing an image cache could adversely effect its ability to be displayed in two or more different charts simultaneously.  In order to prevent that from becoming an issue, all responsibility for creating and maintaining an image cache of any kind would more reasonably fall to the Renderer instance.  UserData as a concept in IDataSet is removed in this change set.