This is a suggestion to fix issue #2083 and maybe also #2070.
The two ArrayList storing a value and its mapping to a formatted string represent a map. Thus, replacing the two ArrayList with a HashMap makes sense.
Furthermore, the runtime behavior of this classes will change from O(n) to O(1) given that a lookup in the hash map is naturally faster than iterating over a list.
The original implementation despite its announcement did not circumvent auto-boxing of the primitive values. The new version reduces the auto-boxing to one.