Enable viewport rect bigger than chart with scale
Enables to set viewport rectangle as bigger than chart range by using
viewport scale.
Feature can be accessed by direct viewport handler manipulation.
In the current code it can be achived with setting mMaxScale < 1, but it
also blocks zomming in/out.
In order to unblock this feature, there is a need for change in
condition for mMinScale as to be possible to be smaller than 1 (but
bigger than 0 -> case 0 is infinite zoom out).

Users who manipulate viewport are expected to know what they are looking
for, so they shouldn't be limited to fixed zoom out value. This way
charts have option to set viewport rect outside chart range and keep
zoom availiability at the same time.