Added the forceRestart animator boolean #2220
Now the chart will not restart when a new animation is setted. It will continue from the point it stopped.
You can force the restart from the beginning by canceling it manualy, or using the forceRestart boolean.

All the previous methods signatures are conserved.
The default behavior is to continue from the point it stopped.