AxisRenderer: Tweaked label interval normalization
Made this slight tweak in a project I was working on today. I had a granularity of 60000 (1 minute), but because of the normalization, labels were still being printed with an interval of 100000. I changed it so if the target interval is below the granularity value, it'll skip the normalization step. This allowed the labels to align with my data points when zoomed all the way in, which were also spaced one every 60000. Probably other ways to achieve that result, and this one won't work for irregularly spaced points or X values that aren't multiples of the granularity, but I don't think this one will cause any further issues in those cases.