demonstrate the problem
In BubbleChartHighlightActivity, click on the bubbles. Some clicks
work properly (highlight the bubble that was clicked), some clicks
highlight the wrong bubble, and some clicks do nothing.

I also refactored MainActivity. Previously it was painful to add
a new activity, especially if it is not at the bottom of the list.
The new implementation is cleaner - WISYWIG and easier to maintain.