refactor saveTouchStart
saveTouchStart no longer runs a search through the data unless there
is at least one inverted axis on the chart.

Previously, saveTouchStart found the nearest entry and saved the data
set that the entry belonged to (as mClosestDataSetToTouch). In some
subsequent operations, e.g. drag, the data set was checked to see
if its axis is inverted.

If there is no inverted axis on the chart, we know a priori that
the check will always be negative. We skip the search in this case.

mClosestDataSetToTouch is replaced with mIsInverted. This does not
change the function, but makes the code much more readable.