optimize highlight processing
A highlight has x and y coordinates. When it is processed, e.g. drawn,
the element associated with the highlight is found by doing a search
through the data values. This search happens repeatedly whenever a
touch event occurs.

This commit uses the Highlight mDataIndex field to remember which
entry the highlight was generated from. Therefore we no longer
need to perform repeated searches.