create OnChartAxisSelectedListener interface
When an axis is selected, the listener can callback to the activity
via the onAxisSelected and onNothingSelected methods.

To enable callbacks, the activity must implement
OnChartAxisSelectedListener and the dataset setHighlightEnabled
must be true.

OnChartAxisSelectedListener was created as a separate class so as to
not break existing code that uses OnChartValueSelectedListener.