cleanup existing code base
Add Legend.setCustom(int[] colors, String[] labels)
This method is documented in the wiki, but does not exist.
Now it does.

Make Axis mAxisMaximum and mAxisMinimum protected,
as is says that should be in the java doc. I
will not be burned again!

Add a method to set the label color. This allows you to
make the default labels invisible (white) ands only
render the highlighted label.