Highlight has type; create Highlights
Highlight now has a Type indicating the source of the highlight,
data value, x axis, left axis right axis or null. The NULL type is
intended to replace a null highlight, ala Kotlin.

Highlights is a collection of Highlight's which avoids
the use of an array (which is hardwired into the legacy
Highlight implementation).