Update LICENSE
The LICENSE file was not properly filled out. It was missing some templates that were supposed to be filled in at the end of the license.  Additionally, the entire Apache 2.0 license is not required on a project that makes use of it. Only this disclaimer is required.

See http://www.apache.org/licenses/LICENSE-2.0#apply under the "How to apply the Apache License to your work" for more information.