Fixed bug #2450 -> Update chart throws Exception
Fixed a IndexOutOfRangeException that occurred irregularly if the chart gets permanently updated.
The bug occurs in the legend renderer. I put the responsible code into a synchronized block and surrounded the critical code by a try-catch. Despite the synchronized, the exception gets sometimes thrown, which the try block catches and replaces the needed vertical space by a cached value.