Fix Drawable LineChart Fill Bug
The LineChart fill draws using intervals to avoid
memory issues, but with a larger data set and many
different (darker) colors Android does not draw the
drawable all the way to the edge of the interval.
This makes it look like there are spaces between the fill.
Draw a little beyond the existing interval so there are no gaps.

https://github.com/PhilJay/MPAndroidChart/pull/5284